feat: add /providers
