+++ lib/providers/index.ts
@@ -1,0 +1,58 @@
+import Claude from "@/components/icons/claude"
+import DeepSeek from "@/components/icons/deepseek"
+import Gemini from "@/components/icons/gemini"
+import Grok from "@/components/icons/grok"
+import Mistral from "@/components/icons/mistral"
+import OpenAI from "@/components/icons/openai"
+import OpenRouter from "@/components/icons/openrouter"
+import Xai from "@/components/icons/xai"
+
+export type Provider = {
+  id: string
+  name: string
+  available: boolean
+  icon: React.ComponentType<React.SVGProps<SVGSVGElement>>
+}
+
+export const PROVIDERS: Provider[] = [
+  {
+    id: "openrouter",
+    name: "OpenRouter",
+    icon: OpenRouter,
+  },
+  {
+    id: "openai",
+    name: "OpenAI",
+    icon: OpenAI,
+  },
+  {
+    id: "mistral",
+    name: "Mistral",
+    icon: Mistral,
+  },
+  {
+    id: "deepseek",
+    name: "DeepSeek",
+    icon: DeepSeek,
+  },
+  {
+    id: "gemini",
+    name: "Gemini",
+    icon: Gemini,
+  },
+  {
+    id: "claude",
+    name: "Claude",
+    icon: Claude,
+  },
+  {
+    id: "grok",
+    name: "Grok",
+    icon: Grok,
+  },
+  {
+    id: "xai",
+    name: "XAI",
+    icon: Xai,
+  },
+] as Provider[]
